chore move editStore to storeRouter
