+++ controllers/seller/editStore.js
@@ -1,6 +1,6 @@
 const Store = require('../../models/store')
 
-const { validateEditInput } = require('../../middlewares/validateEditInput')
+const { validateEditInput } = require('../../middleware/validateEditInput')
 
 function editStore(req, res) {
   const { sub: sellerId } = req.decodedToken
@@ -26,13 +26,11 @@ function editStore(req, res) {
         storeName
       }
 
-      Store.findOneAndUpdate(
-        sellerId,
-        { $set: newStoreDetails },
-        { new: true }
-      ).then(newStore => {
-        res.json(newStore)
-      })
+      Store.findOneAndUpdate(sellerId, { $set: newStoreDetails }, { new: true })
+        .then(newStore => {
+          res.json(newStore)
+        })
+        .catch(err => console.log(err))
     })
     .catch(err => {
       console.log(err)
